+++ ephem_read.h
@@ -27,24 +27,24 @@
 /*  Interpolate_Libration                                                     */
 /*----------------------------------------------------------------------------*/
 
-    void Interpolate_Libration( double Time, int Target, double Libration[3] );
+    void Interpolate_Libration( long double Time, int Target, long double Libration[3] );
 
 /*----------------------------------------------------------------------------*/
 /*  Interpolate_Nutation                                                      */
 /*----------------------------------------------------------------------------*/
 
-    void Interpolate_Nutation( double Time , int Target , double Nutation[2] );
+    void Interpolate_Nutation( long double Time , int Target , long double Nutation[2] );
    
 /*----------------------------------------------------------------------------*/
 /*  Interpolate_Position                                                      */
 /*----------------------------------------------------------------------------*/
 
-    void Interpolate_Position( double Time , int Target , double Position[3] );
+    void Interpolate_Position( long double Time , int Target , long double Position[3] );
 
 /*----------------------------------------------------------------------------*/
 /*  Interpolate_State                                                         */
 /*----------------------------------------------------------------------------*/
 
-    void Interpolate_State( double Time , int Target , stateType *Planet );
+    void Interpolate_State( long double Time , int Target , stateType *Planet );
 
 double Find_DataInHeader( char    *szTarget);
